Create PayoutDetailsPersonalDetails subcomponent
